Add many new features from incidents database
Add features for the number of suicides, incidents involving juveniles,
domestic violence incidents, hate events, narcotics-related incidents,
gang-related incidents the officer has dealt with in the last year.
Features for the number of gun/knife-related incidents and personal
weapons (fists/feet/teeth) incidents the officer has dealt with in the
last year. Also add the average and minimum age of victims the officer
has dealt with in the last year.